+++ eslint.config.mjs
@@ -239,11 +239,7 @@ export default tseslint.config(
         ...globals.browser,
       },
     },
-    extends: [
-      ...sharedExtends,
-      eslintReactRefresh.configs.vite,
-      internalPlugin.configs.importRules,
-    ],
+    extends: [...sharedExtends, eslintReactRefresh.configs.vite],
     plugins: {
       ...sharedPlugins,
     },
@@ -254,7 +250,7 @@ export default tseslint.config(
   {
     files: ['**/*.{ts,tsx}'],
     settings: sharedSettings,
-    extends: [...sharedExtends, ...typescriptExtends, internalPlugin.configs.allRules],
+    extends: [...sharedExtends, ...typescriptExtends],
     plugins: {
       ...sharedPlugins,
       ...typescriptPlugins,
@@ -262,7 +258,6 @@ export default tseslint.config(
     rules: {
       ...typescriptRules,
       ...sharedRules,
-      // 'internal/safely-spread-props': 'warn',
     },
   },
   {
@@ -280,7 +275,7 @@ export default tseslint.config(
       'packages/**/*.spec.*',
     ],
     settings: sharedSettings,
-    extends: [...sharedExtends, ...typescriptExtends, internalPlugin.configs.allRules],
+    extends: [...sharedExtends, ...typescriptExtends],
     plugins: {
       ...sharedPlugins,
       ...typescriptPlugins,
@@ -289,13 +284,12 @@ export default tseslint.config(
       ...sharedRules,
       ...typescriptRules,
       ...packageProductionRules,
-      // 'internal/safely-spread-props': 'warn',
     },
   },
   {
     files: ['**/*mobile*/**/*.{ts,tsx}'],
     settings: sharedSettings,
-    extends: [...sharedExtends, ...typescriptExtends, internalPlugin.configs.importRules],
+    extends: [...sharedExtends, ...typescriptExtends],
     plugins: {
       ...sharedPlugins,
       ...typescriptPlugins,
@@ -305,6 +299,11 @@ export default tseslint.config(
       ...sharedRules,
       ...typescriptRules,
     },
+  },
+  // Rules specific to mobile story files
+  {
+    files: ['packages/mobile/**/*.stories.tsx'],
+    extends: [internalPlugin.configs.mobileStoryRules],
   },
   {
     files: ['**/*.test.{ts,tsx}', '**/__tests__/**', '**/setup.js'],
@@ -319,7 +318,6 @@ export default tseslint.config(
       ...typescriptExtends,
       eslintJest.configs['flat/recommended'],
       eslintTestingLibrary.configs['flat/react'],
-      internalPlugin.configs.importRules,
     ],
     plugins: {
       ...sharedPlugins,