+++ libs/eslint-plugin-internal/src/index.mjs
@@ -1,40 +1,18 @@
+import exampleScreenContainsExampleRule from './example-screen-contains-example/index.mjs';
+import exampleScreenDefaultRule from './example-screen-default/index.mjs';
 import safelySpreadPropsRule from './safely-spread-props/index.mjs';
-import importAutofixRule from './import-autofix.mjs';
 
 const plugin = {
   name: '@coinbase/eslint-plugin-internal',
   rules: {
-    'import-autofix': importAutofixRule,
     'safely-spread-props': safelySpreadPropsRule,
+    'example-screen-default': exampleScreenDefaultRule,
+    'example-screen-contains-example': exampleScreenContainsExampleRule,
   },
   configs: {},
 };
 
 Object.assign(plugin.configs, {
-  allRules: {
-    plugins: {
-      internal: plugin,
-    },
-    languageOptions: {
-      parserOptions: {
-        projectService: true,
-        tsconfigRootDir: import.meta.dirname,
-      },
-    },
-    rules: {
-      'internal/import-autofix': 'error',
-      // 'internal/safely-spread-props': ['error', { maxInvalidPropsInMessage: 5 }],
-    },
-  },
-  importRules: {
-    plugins: {
-      internal: plugin,
-    },
-    rules: {
-      'internal/import-autofix': 'error',
-      // 'internal/safely-spread-props': 'off',
-    },
-  },
   typedRules: {
     plugins: {
       internal: plugin,
@@ -46,8 +24,16 @@ Object.assign(plugin.configs, {
       },
     },
     rules: {
-      'internal/import-autofix': 'off',
       // 'internal/safely-spread-props': ['error', { maxInvalidPropsInMessage: 5 }],
+    },
+  },
+  mobileStoryRules: {
+    plugins: {
+      internal: plugin,
+    },
+    rules: {
+      'internal/example-screen-default': 'warn',
+      'internal/example-screen-contains-example': 'warn',
     },
   },
 });