+++ packages/mobile/src/page/__stories__/PageHeaderInPage.stories.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
-import { View } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { NoopFn } from '@coinbase/cds-common/utils/mockUtils';
 
 import { Button, ButtonGroup, IconButton } from '../../buttons';
-import { ExampleScreen } from '../../examples/ExampleScreen';
-import { VStack } from '../../layout';
+import { Example, ExampleScreen } from '../../examples/ExampleScreen';
+import { Box, VStack } from '../../layout';
 import { Text } from '../../typography/Text';
 import { PageFooter } from '../PageFooter';
 import { PageHeader } from '../PageHeader';
@@ -24,8 +23,8 @@ const PageHeaderInPage = () => {
   const inset = useSafeAreaInsets();
 
   return (
-    <View>
-      <ExampleScreen>
+    <ExampleScreen paddingX={0}>
+      <Example hideDivider height="100%">
         <VStack gap={0} width="100%">
           <PageHeader
             background="bg"
@@ -48,11 +47,11 @@ const PageHeaderInPage = () => {
             <Text font="title1">Primary Content</Text>
           </VStack>
         </VStack>
-      </ExampleScreen>
-      <View style={{ position: 'absolute', bottom: inset.bottom / 2, left: 0, right: 0 }}>
-        <PageFooter action={exampleProps.endButtons} />
-      </View>
-    </View>
+        <Box style={{ position: 'absolute', bottom: inset.bottom / 2, left: 0, right: 0 }}>
+          <PageFooter action={exampleProps.endButtons} />
+        </Box>
+      </Example>
+    </ExampleScreen>
   );
 };